+++ CentroLink/CentroLink/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## 03/05/2021
 - 51728 - Centrolink - Create View in to load Ticket schedules
+- 51730 - Centrolink - Create View to edit Ticket schedules
 
 ### fixes
 - Adjusted copy on buttons from 'PromoTicket' to 'Promo Ticket'